# Conflicts:
#	VERSION

+++ bandwidthrtc/src/main/kotlin/com/bandwidth/rtc/signaling/rpc/SDPOfferNotification.kt
@@ -8,6 +8,8 @@ data class SDPOfferNotification(
     val endpointId: String? = null,
     val peerType: String? = null,
     val sdpOffer: String,
-    val sdpRevision: Int? = null,
+    // A signaling-side revision id, not a small sequence counter - the gateway sends values
+    // large enough to overflow Int (observed: millisecond timestamps), so this must stay Long.
+    val sdpRevision: Long? = null,
     val trackMetadata: Map<String, TrackMetadata>? = null
 )
