+++ bandwidthrtc/src/main/kotlin/com/bandwidth/rtc/webrtc/PeerConnectionManagerInterface.kt
@@ -26,7 +26,7 @@ interface PeerConnectionManagerInterface {
     suspend fun applyPublishAnswer(remoteAnswer: String)
     suspend fun handleSubscribeSdpOffer(
         sdpOffer: String,
-        sdpRevision: Int?,
+        sdpRevision: Long?,
         metadata: Map<String, TrackMetadata>?
     ): String
